pyproject.toml
[project]
name = "ratp-calendrier-travaux"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "icalendar>=6.1.1",
    "ipykernel>=6.29.5",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.0.1",
    "pyvirtualdisplay>=3.0",
    "requests>=2.32.3",
    "seleniumbase>=4.34.1",
]
